@@ -13,7 +13,7 @@ import java.util.function.Supplier;
 public final class RoleConfig {
     public static final Codec<RoleConfig> CODEC = RecordCodecBuilder.create(i -> i.group(
             Codec.intRange(0, Integer.MAX_VALUE).optionalFieldOf("level", 0).forGetter(c -> c.level),
-            MoreCodecs.propagatingOptionalFieldOf(RoleOverrideMap.CODEC, "overrides", (Supplier<RoleOverrideMap>) RoleOverrideMap::new).forGetter(c -> c.overrides),
+            RoleOverrideMap.CODEC.optionalFieldOf("overrides", new RoleOverrideMap()).forGetter(c -> c.overrides),
             MoreCodecs.arrayOrUnit(Codec.STRING, String[]::new).optionalFieldOf("includes", new String[0]).forGetter(c -> c.includes),
             RoleApplyConfig.CODEC.optionalFieldOf("apply").forGetter(c -> Optional.ofNullable(c.apply))
     ).apply(i, RoleConfig::new));
@@ -25,7 +25,7 @@ public final class RoleConfig {
 
     private RoleConfig(int level, RoleOverrideMap overrides, String[] includes, Optional<RoleApplyConfig> apply) {
         this.level = level;
-        this.overrides = overrides;
+        this.overrides = new RoleOverrideMap(overrides);
         this.includes = includes;
         this.apply = apply.orElse(null);
     }
