@@ -5,6 +5,7 @@ import dev.gegy.roles.store.PlayerRoleManager;
 import net.fabricmc.fabric.api.util.NbtType;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.nbt.NbtCompound;
+import net.minecraft.nbt.NbtElement;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
@@ -21,8 +22,8 @@ public abstract class ServerPlayerEntityMixin extends PlayerEntity {
 
     @Inject(method = "readCustomDataFromNbt", at = @At("RETURN"))
     private void readCustomDataFromTag(NbtCompound tag, CallbackInfo ci) {
-        if (tag.contains("roles", NbtType.LIST)) {
-            PlayerRoleManager.get().addLegacyRoles((ServerPlayerEntity) (Object) this, tag.getList("roles", NbtType.STRING));
+        if (tag.contains("roles", NbtElement.LIST_TYPE)) {
+            PlayerRoleManager.get().addLegacyRoles((ServerPlayerEntity) (Object) this, tag.getList("roles", NbtElement.STRING_TYPE));
         }
     }
 }
